Adds Python module to build Language ADT
Sourcing data from Wikipedia, we're building a list of known
languages and using them as constructors for a `Language` ADT that
will serve as the bedrock of the Ogma system. The hope is that
they can be used as type tags to serve a type family that will
restrict permissible values in a BCP-47 language code. This ADT
itself will also be able to provide information about the language
such as the name, linguonyms, ISO 639 codes, scripts, and more.

In order to construct this ADT, a Python program is provided that
scrapes Wikipedia for the data. Currently, it only finds the data
and builds the stated Haskell module. Future steps will include
adding to-bytes and to-text functions for the name, and scraping
more data for more of the same. This script has been set up to run
asynchronously to ease a future pain point of scraping the Wiki
page for each of these `Language` constructors to get additional
data. It is expected that once this script is able to scrape all
of the data needed, it will not have to be updated very often.