Separates language_maker, now Babilu, into separate modules
This renames `language_maker`, a rather uninspired name, into
`babilu`, for the Sumerian tower upon which the Tower of Babel was
inspired.

Soon we will be adding additional steps, such as building a module
for a language script ADT. Apart from the rename, this is just
organizing for the future.