Adds local caching of scraped web pages for faster dev loop
This locally saves the `table` elements from each web page scraped
upon finding them for the first time. This will drastically speed
up dev time as we will no longer have to wait for the pages to be
fetched (even though it was done asynchronously). It should also
speed up parsing time since we're only saving the table (which will
save a lot of space once we get to individual script and language
pages).