Re-organizes Ogma, removes old generated code
Since we are pivoting to using the IANA subtag registry to generate
code for Ogma's BCP-47 compliance, we needed a change to the way
Ogma was structured. This gets rid of the Python web scraping code
and the Haskell code it generated, and instead adds a new Haskell
executable `ogma-gen` that will parse the `data/` files and build
new Haskell code, which it will deposit into the main library. This
also updates some of the `./scripts` to account for this change.

The subtag registry file was added so the compiler would stop
complaining about the `data/` directory being empty, but the
code generator will likely fetch that data each time it's run to
ensure it's always up to date.