Generates Language type from IANA subtag registry
This adds an executable that fetches the latest version of the
IANA subtag registry and parses it for languages. Each language is
then turned into a constructor for a `Language` ADT, converted to
name bytes and text, and convert to and from ISO 639 code bytes
and text; all this is done in generated code produced by the new
executable.

This fundamentally returns us to the previous state of Ogma, but
with a more reliable and accurate source for BCP-47 supported
language data.